requirements.txt
ffmpeg_python==0.2.0
mutagen==1.47.0
numpy==2.2.3
Pillow==11.1.0
psutil==7.0.0
pycdlib==1.14.0
pyzbar==0.1.9
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
colorama==0.4.6
ConfigParser==7.1.0
cryptography==44.0.0
docutils==0.21.2
ffmpeg-python
filelock==3.17.0
h2==4.2.0
HTMLParser==0.0.2
ipython==8.32.0
ipywidgets==8.1.5
js==1.0
keyring==25.6.0
mutagen==1.47.0
Pillow==11.1.0
protobuf==5.29.3
python-slugify==8.0.4
redis==5.2.1
simplejson==3.19.3
Sphinx==8.1.3
thread==2.0.5
zstandard==0.23.0
opencv-python
tk
platformdirs